+++ firmware/hdl/data/gtl/transverse_mass.vhd
@@ -19,6 +19,7 @@ entity transverse_mass is
         N_OBJ_1 : positive;
         N_OBJ_2 : positive;
         OBJ : obj_type_array;
+        BX : bx_array;
         PT1_WIDTH : positive;
         PT2_WIDTH : positive;
         COSH_COS_WIDTH : positive;
@@ -47,7 +48,7 @@ begin
             conv_i: for k in 0 to  COSH_COS_WIDTH-1 generate
                 cos_dphi_i(i,j)(k) <= cos_dphi(i,j,k);
             end generate conv_i;
-            same_obj_t: if (OBJ(1) = OBJ(2)) and j>i generate
+            same_obj_t: if (OBJ(1) = OBJ(2)) and (BX(1) = BX(2)) and j>i generate
 -- less resources
                 mass_calc_i : entity work.trans_mass_calc
                     generic map(PT1_WIDTH, PT2_WIDTH, COSH_COS_WIDTH, COSH_COS_PREC, MASS_WIDTH)  
@@ -59,7 +60,7 @@ begin
                     transverse_mass_sq_div2(i,j) <= transverse_mass_sq_div2_temp(i,j); 
                     transverse_mass_sq_div2(j,i) <= transverse_mass_sq_div2_temp(i,j); 
             end generate same_obj_t;    
-            diff_obj_t: if (OBJ(1) /= OBJ(2)) generate
+            diff_obj_t: if (OBJ(1) /= OBJ(2)) or ((OBJ(1) = OBJ(2)) and (BX(1) /= BX(2))) generate
                 mass_calc_i : entity work.trans_mass_calc
                     generic map(PT1_WIDTH, PT2_WIDTH, COSH_COS_WIDTH, COSH_COS_PREC, MASS_WIDTH)  
                     port map(